cleanup derivative utility
